requirements.txt
python-telegram-bot~=12.6.1
SQLAlchemy~=1.3.16
psycopg2~=2.8.5
